requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
attrs==21.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
django-rest-passwordreset==1.3.0
djangorestframework==3.12.4
dlib @ file:///dlib-19.22.99-cp310-cp310-win_amd64.whl
drf-spectacular==0.17.3
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
keras==2.10.0
keras-facenet==0.3.2
Keras-Preprocessing==1.1.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
mtcnn==0.1.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
Pillow==9.3.0
protobuf==3.19.6
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.19.3
python-decouple==3.4
pytz==2021.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.15
Werkzeug==2.2.3
whitenoise==5.3.0
wrapt==1.15.0
waitress==2.1.2
channels==4.0.0